first pass of reorganizing - util.py split up
